refactor: queries ts
