+++ apps/client/app/employee/[id]/report/[cycleId]/page.tsx
@@ -1,14 +1,14 @@
 import { apolloClient } from '@/lib/client'
 import '../../../../global.css'
-import { GET_REPORT_FOR_EMPLOYEE, GET_USER_FULLNAME_BY_ID } from '@/lib/queries'
+import { queries } from '@/lib/graphql-queries'
 
 
 async function Report({ params }) {
 
   const variables = { targetId: params.id, cycleId: params.cycleId }
 
-  const { data: { getUser: { fullName } } } = await apolloClient.query({ query: GET_USER_FULLNAME_BY_ID, variables: { id: params.id } })
-  const { data: { getReport } } = await apolloClient.query({ query: GET_REPORT_FOR_EMPLOYEE, variables })
+  const { data: { getUser: { fullName } } } = await apolloClient.query({ query: queries.GET_USER_FULLNAME_BY_ID, variables: { id: params.id } })
+  const { data: { getReport } } = await apolloClient.query({ query: queries.GET_REPORT_FOR_EMPLOYEE, variables })
 
   return (
     <div className='p-4'>
